iconclass_extraction/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
blis==0.7.11
catalogue==2.0.10
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
cloudpathlib==0.16.0
confection==0.1.3
cymem==2.0.8
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0-py3-none-any.whl
fastapi==0.104.1
h11==0.14.0
idna==3.4
Jinja2==3.1.2
langcodes==3.3.0
MarkupSafe==2.1.3
murmurhash==1.0.10
numpy==1.26.1
packaging==23.2
preshed==3.0.9
pydantic==2.4.2
pydantic_core==2.10.1
python-dotenv==1.0.0
requests==2.31.0
smart-open==6.4.0
sniffio==1.3.0
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
starlette==0.27.0
thinc==8.2.1
tqdm==4.66.1
typer==0.9.0
typing_extensions==4.8.0
urllib3==2.0.7
uvicorn==0.23.2
wasabi==1.1.2
weasel==0.3.3
